requirements.txt
numpy
langchain
langchain_groq
langchain_core
langchain_community
psycopg2-binary